[#166] refactor: Car infra 코드 파일위치 변경
