+++ rest/src/main/java/com/wherecar/rest/car/infrastructure/CarStore.java
@@ -1,4 +1,4 @@
-package com.wherecar.rest.car.infrastructure.infra;
+package com.wherecar.rest.car.infrastructure;
 
 import com.wherecar.rest.car.domain.Car;
